requirements.txt
# Bordro Motoru - Gereksinimler
pywebview>=6.0
pdfplumber>=0.11
Flask>=3.0
